refactor: share remote config parsing primitives
