+++ src/remote-config.ts
@@ -2,5 +2,5 @@ export type {
   RemoteConfigProfile,
   RemoteConfigProfileOptions,
   ResolvedRemoteConfigProfile,
-} from './remote-config-core.ts';
+} from './remote-config-schema.ts';
 export { resolveRemoteConfigPath, resolveRemoteConfigProfile } from './remote-config-core.ts';
